MMCA 4786 - Upgrade scala & library dependancies to latest version (#136)
* MMCA-4786 : initial commit

* MMCA-4786 : fixed connectors,  app dependencies

* MMCA-4786 : fixed urls, error handler, layout

* MMCA-4786 : fixed test cases - url, imports of mockito, when and any

* MMCA-4786 : fixed formatting warnings, added shouldBe and removed shouldWrapper

* MMCA-4786 : fixed mocks in connectors spec

* MMCA-4786 : fixed platops bot comments and cleaned up connector specs

* MMCA-4786 : fixed TODOs and sdes spec

* MMCA-4786 : removed a TODO as it is unnecessary

* MMCA-4786 : fixed scalastyle warnings and removed default scala settings

* MMCA-4786 : removed exhaust depth settings and its being ignored during compilation

* MMCA-4786 : formatted the  files

* MMCA-4786 : removed redundant lines

* MMCA-4786: added layout spec and removed redundant case

* MMCA-4786: PR comment fixes 01

* MMCA-4786: PR comment fixes 02

* MMCA-4786 : updated frontend-hmrc-play version

+++ app/repositories/SessionRepository.scala
@@ -20,6 +20,8 @@ import com.mongodb.client.model.Indexes.ascending
 import models.UserAnswers
 import org.mongodb.scala.model.Filters.equal
 import org.mongodb.scala.model.{IndexModel, IndexOptions, ReplaceOptions}
+import org.mongodb.scala.SingleObservableFuture
+import org.mongodb.scala.ToSingleObservablePublisher
 import play.api.Configuration
 import uk.gov.hmrc.mongo.play.PlayMongoComponent
 import uk.gov.hmrc.mongo.play.json.PlayMongoRepository
@@ -41,7 +43,7 @@ class DefaultSessionRepository @Inject()(mongoComponent: PlayMongoComponent, con
       IndexModel(
         ascending("lastUpdated"),
         IndexOptions().name("user-answers-last-updated-index")
-          .expireAfter(config.get[Int]("mongodb.timeToLiveInSeconds"), TimeUnit.SECONDS)
+          .expireAfter(config.get[Long]("mongodb.timeToLiveInSeconds"), TimeUnit.SECONDS)
       )
     )
   ) with SessionRepository {
@@ -53,9 +55,10 @@ class DefaultSessionRepository @Inject()(mongoComponent: PlayMongoComponent, con
 
   override def set(userAnswers: UserAnswers): Future[Boolean] =
     collection.replaceOne(
-      equal("_id", userAnswers.id),
-      userAnswers.copy(lastUpdated = LocalDateTime.now()),
-      ReplaceOptions().upsert(true))
+        equal("_id", userAnswers.id),
+        userAnswers.copy(lastUpdated = LocalDateTime.now()),
+        ReplaceOptions().upsert(true)
+      )
       .toFuture()
       .map(_.wasAcknowledged())
